server/requirements.txt
# Briefly Cloud - Python Dependencies
# Optimized for Python 3.11.9

# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==24.1.0
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.0

# Supabase client
supabase==2.3.4

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Stripe for payments
stripe==8.2.0

# Scientific computing
numpy==1.25.2
scipy==1.11.4
scikit-learn==1.3.2

# Vector store and ML - Fixed versions for compatibility
chromadb==0.4.15
sentence-transformers==2.2.2
transformers==4.35.2
huggingface-hub==0.17.3

# Additional utilities
python-dateutil==2.8.2
requests==2.31.0

# Document parsing libraries
pdfplumber==0.10.3
python-docx==1.1.0
openpyxl==3.1.2
xlrd==2.0.1
python-pptx==0.6.23
striprtf==0.0.26
odfpy==1.4.1
ebooklib==0.18
pyspellchecker==0.8.1
# XML handled by Python stdlib

